move CameraInfo to dto namespace
